fix a race when kafka and zk are starting (and restarting) at the same time
Summary:
we observed the following error on the broker

```
[2018-07-06 18:28:02,112] FATAL Fatal error during KafkaServerStartable startup. Prepare to shutdown (kafka.server.KafkaServerStartable)
java.lang.RuntimeException: A broker is already registered on the path /brokers/ids/0. This probably indicates that you either have configured a brokerid that is already in use, or else you have shutdown this broker and restarted it faster than the zookeeper timeout so it appears to be re-registering.
    at kafka.utils.ZkUtils.registerBrokerInZk(ZkUtils.scala:408)
    at kafka.utils.ZkUtils.registerBrokerInZk(ZkUtils.scala:394)
    at kafka.server.KafkaHealthcheck.register(KafkaHealthcheck.scala:71)
    at kafka.server.KafkaHealthcheck.startup(KafkaHealthcheck.scala:51)
    at kafka.server.KafkaServer.startup(KafkaServer.scala:269)
    at kafka.server.KafkaServerStartable.startup(KafkaServerStartable.scala:39)
    at kafka.Kafka$.main(Kafka.scala:67)
    at kafka.Kafka.main(Kafka.scala)
```

it is caused by a kafka broker process still keeping a connection to zookeeper.  if supervisord starts a new broker, the /borker/ids path in zk has a race
between old connection being dropped and new one being created.  this change tries to stop the kafka broker before starting a new one.

this change also extends the expiration of SSL certificates on the test container

Test Plan:
psyduck kerberos_sanity.py
http://psyduck.memcompute.com/run/fd62674d0ad0426bb11ae6b3b7c75f6e/tests/all
http://psyduck.memcompute.com/run/43917ab8a2164c24b9a371f413643f34/tests/all

Reviewers: tfly, adam, boutin, joseph

Reviewed By: joseph

Subscribers: engineering-list

JIRA Issues: DB-33026

Differential Revision: https://grizzly.memsql.com/D27176