Refactoring certificate creation
Summary:
SSL fips tests are failing using the current certificate because it is 1048 bits while
fips requires >2048 bits. As a part of debugging this I rewrote the certificate creation to
simplify. Firstly the public and private keys are create through openssl and then imported to
keytool. Secondly instead of creating a CA cert and then signing the x509 with the CA cert,
we will just use a self-signed server cert.

Test Plan:
.pipelines
http://psyduck.memcompute.com/run/e2fd44c425844701b27d5321a1a745e1/tests/all
.pipelines_fips
http://psyduck.memcompute.com/run/da9fd06c79e041c7a7d9801c6463d4ca/tests/all

Reviewers: sasha, mkobyakov, eli

Reviewed By: mkobyakov

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D46306