Adding script to build all supported docker versions and fixes
Summary:
This change adds a script to build and push the supported docker images. This is needed because the
aio containers use 1024bit RSA certs which dont work in fips.

Additionally, the python-twitter has bubmped downstream dependencies which use f-strings which
isn't suppported in the python3.5 downloaded. I went down the rabithole of downlaoding and building
python3.7 but I dont think any test actually depends on the package so I am removing it + the
pyhton install.

Additionally, I am moving the schema cert to a new hostname, host2.example.com, so that it doesn't
conflict with the kafka cert if we are running schema+ssl and kafka+ssl on different containers.

Test Plan:
.pipelines
http://psyduck.memcompute.com/run/0337bd12dd224af0af2e0be398ac21e7/tests/all
.pipelines_fips
http://psyduck.memcompute.com/run/19488582e9bd4285afafe83578c2eb35/tests/all

Reviewers: sasha, mkobyakov

Reviewed By: mkobyakov

Subscribers: mmikhno-ua, jhuang, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D46747